feat: add react-web-hocs, refactor withLoading
