+++ README.md
@@ -102,7 +102,7 @@ $ yarn analyze
 - [x] Set up i18n
 - [x] Set up React router
 - [x] Set up code splitting
-- [ ] Set up basics HOCs: loading, updating, error,...
+- [x] Set up basics HOCs: loading, updating, error,...
 - [x] Set up common utilities: common, date, theme,...
 - [x] Set up emotion, tailwind for UI components
 - [x] Set up storybook
