+++ package.json
@@ -47,6 +47,7 @@
     "react-loadable": "5.5.0",
     "react-redux": "5.0.7",
     "react-router-dom": "4.3.1",
+    "react-web-hocs": "0.0.2",
     "recompose": "0.30.0",
     "redux": "4.0.1",
     "redux-actions": "2.6.1",
